+++ lib/thirdparty/flutter_context_menu/components/menu_item.dart
@@ -35,6 +35,7 @@ final class MenuItem<T> extends ContextMenuItem<T> {
   final String label;
   final IconData? icon;
   final double? iconOpacity;
+  final Widget? leading;
   final BoxConstraints? constraints;
   final bool enabled;
   final TextStyle? textStyle;
@@ -44,6 +45,7 @@ final class MenuItem<T> extends ContextMenuItem<T> {
     required this.label,
     this.icon,
     this.iconOpacity,
+    this.leading,
     super.value,
     super.onSelected,
     super.keepMenuOpen,
@@ -58,6 +60,7 @@ final class MenuItem<T> extends ContextMenuItem<T> {
     required List<ContextMenuEntry> items,
     this.icon,
     this.iconOpacity,
+    this.leading,
     super.onSelected,
     super.keepMenuOpen,
     this.constraints,
@@ -110,13 +113,14 @@ final class MenuItem<T> extends ContextMenuItem<T> {
                     SizedBox(width: 4.0),
                     SizedBox.square(
                       dimension: 32.0,
-                      child: Icon(
-                        icon,
-                        size: 16.0,
-                        color: foregroundColor.withOpacity(
-                          iconOpacity ?? foregroundColor.a,
-                        ),
-                      ),
+                      child: leading ??
+                          Icon(
+                            icon,
+                            size: 16.0,
+                            color: foregroundColor.withOpacity(
+                              iconOpacity ?? foregroundColor.a,
+                            ),
+                          ),
                     ),
                     if (label.isNotEmpty) const SizedBox(width: 4.0),
                     Expanded(